Completed.
Unit Tests include:
posittive test cases and negative test case, also tests the internal library.

Console Application include
takes in the items to the collection and then displays them out.

Library defines 3 classes with the receipt collection implementing the ILIST intefrace to allow the sum value of the collection to be update upon change of the collection.
This could of been done a bit diffrent with the INotify event hookups.